Overhaul receiving rate
I worked some changes to the functions that fetch the messages from gmail to allow it to automatically delete the messages after harvesting the data from them. This allows a smaller search index when searching for the latest email (i.e the only one) and thus grants incredible speed at reading the gmail inbox.